# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health"
        ]
      }
    },
    "/api/admin/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/admin/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "sid": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/admin/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sid": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/admin/templates": {
      "post": {
        "operationId": "TemplatesController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "sid": []
          }
        ],
        "tags": [
          "templates"
        ]
      },
      "get": {
        "operationId": "TemplatesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sid": []
          }
        ],
        "tags": [
          "templates"
        ]
      }
    },
    "/api/admin/templates/{id}": {
      "get": {
        "operationId": "TemplatesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sid": []
          }
        ],
        "tags": [
          "templates"
        ]
      }
    },
    "/api/admin/campaigns": {
      "post": {
        "operationId": "CampaignsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "sid": []
          }
        ],
        "tags": [
          "campaigns"
        ]
      },
      "get": {
        "operationId": "CampaignsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sid": []
          }
        ],
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/admin/campaigns/{id}": {
      "get": {
        "operationId": "CampaignsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sid": []
          }
        ],
        "tags": [
          "campaigns"
        ]
      },
      "patch": {
        "operationId": "CampaignsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sid": []
          }
        ],
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/admin/campaigns/{id}/stats": {
      "get": {
        "operationId": "CampaignsController_stats",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sid": []
          }
        ],
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/admin/analytics/channels": {
      "get": {
        "operationId": "AnalyticsController_channels",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sid": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/admin/analytics/campaigns": {
      "get": {
        "operationId": "AnalyticsController_campaigns",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sid": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/admin/forms": {
      "post": {
        "operationId": "FormsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFormDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "sid": []
          }
        ],
        "tags": [
          "forms"
        ]
      },
      "get": {
        "operationId": "FormsController_findAll",
        "parameters": [
          {
            "name": "campaignId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sid": []
          }
        ],
        "tags": [
          "forms"
        ]
      }
    },
    "/api/admin/forms/{id}": {
      "get": {
        "operationId": "FormsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sid": []
          }
        ],
        "tags": [
          "forms"
        ]
      },
      "patch": {
        "operationId": "FormsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFormDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sid": []
          }
        ],
        "tags": [
          "forms"
        ]
      }
    },
    "/api/admin/forms/{formId}/links": {
      "post": {
        "operationId": "LinksController_create",
        "parameters": [
          {
            "name": "formId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLinkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "sid": []
          }
        ],
        "tags": [
          "links"
        ]
      },
      "get": {
        "operationId": "LinksController_findAll",
        "parameters": [
          {
            "name": "formId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sid": []
          }
        ],
        "tags": [
          "links"
        ]
      }
    },
    "/p/{slug}": {
      "get": {
        "operationId": "PublicPageController_page",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public"
        ]
      }
    },
    "/api/public/forms/{slug}/submissions": {
      "post": {
        "operationId": "PublicApiController_submit",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "public"
        ]
      },
      "options": {
        "operationId": "PublicApiController_options",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "public"
        ]
      }
    },
    "/api/admin/submissions": {
      "get": {
        "operationId": "SubmissionsController_list",
        "parameters": [
          {
            "name": "campaignId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "formId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sid": []
          }
        ],
        "tags": [
          "submissions"
        ]
      }
    }
  },
  "info": {
    "title": "리드마그넷 CRM API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "sid": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sid"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCampaignDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateCampaignDto": {
        "type": "object",
        "properties": {}
      },
      "CreateFormDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateFormDto": {
        "type": "object",
        "properties": {}
      },
      "CreateLinkDto": {
        "type": "object",
        "properties": {}
      },
      "CreateSubmissionDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}